+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 * Require Dart 3.0.
 * Mark mixin classes as `mixin class`.
+* Added `Map.pairs` and `Map.addPairs` extension methods to make it easier to
+  work with maps using Dart 3 record types.
+* Added a `mapFromPairs` function to create a map from Dart 3 record types.
 
 ## 1.17.2